+++ rollup.config.js
@@ -46,6 +46,7 @@ function getRollupObject ({minifying, format = 'umd'} = {}) {
     return nonMinified;
 }
 
+// eslint-disable-next-line import/no-anonymous-default-export
 export default [
     getRollupObject({minifying: true, format: 'umd'}),
     getRollupObject({minifying: false, format: 'umd'}),